refactor: 열람실 API와 학과별 랭킹 API 분리
