+++ src/test/java/com/hongik/controller/study/StudySessionControllerTest.java
@@ -58,7 +58,7 @@ class StudySessionControllerTest {
                 .build();
 
         // when // then
-        mockMvc.perform(post("/api/v1/study/start").with(csrf())
+        mockMvc.perform(post("/api/v1/study/session/start").with(csrf())
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(request))
                 )
@@ -76,7 +76,7 @@ class StudySessionControllerTest {
                 .build();
 
         // when // then
-        mockMvc.perform(patch("/api/v1/study/end").with(csrf())
+        mockMvc.perform(patch("/api/v1/study/session/end").with(csrf())
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(request)))
                 .andDo(print())
@@ -93,7 +93,7 @@ class StudySessionControllerTest {
                 .build();
 
         // when // then
-        mockMvc.perform(post("/api/v1/study/start").with(csrf())
+        mockMvc.perform(post("/api/v1/study/session/start").with(csrf())
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(request)))
                 .andDo(print())
@@ -113,7 +113,7 @@ class StudySessionControllerTest {
                 .build();
 
         // when // then
-        mockMvc.perform(patch("/api/v1/study/end").with(csrf())
+        mockMvc.perform(patch("/api/v1/study/session/end").with(csrf())
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(request)))
                 .andDo(print())
@@ -133,7 +133,7 @@ class StudySessionControllerTest {
                 .build();
 
         // when // then
-        mockMvc.perform(patch("/api/v1/study/end").with(csrf())
+        mockMvc.perform(patch("/api/v1/study/session/end").with(csrf())
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(request)))
                 .andDo(print())
@@ -162,89 +162,12 @@ class StudySessionControllerTest {
                 .willReturn(response);
 
         // when // then
-        mockMvc.perform(get("/api/v1/study").with(csrf())
+        mockMvc.perform(get("/api/v1/study/session").with(csrf())
                         .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.data.length()").value(2))
                 .andExpect(jsonPath("$.data[0].userId").value(1L))
                 .andExpect(jsonPath("$.data[0].nickname").value("김철수"));
-    }
-
-    @DisplayName("특정 날짜에 대한 공부 시간 조회")
-    @Test
-    void getStudyDuration() throws Exception {
-        // given
-        StudyDurationResponse result = StudyDurationResponse.builder().build();
-        LocalDate now = LocalDate.of(2024, 10, 1);
-
-        BDDMockito.given(studySessionService.getStudyDuration(now, 1L))
-                .willReturn(result);
-
-        // when // then
-        mockMvc.perform(
-                        get("/api/v1/study/duration").with(csrf())
-                                .contentType(MediaType.APPLICATION_JSON)
-                )
-                .andDo(print())
-                .andExpect(status().isOk());
-    }
-
-    @DisplayName("특정 연도와 월에 대한 공부 횟수 조회")
-    @Test
-    void getStudyCount() throws Exception {
-        // given
-        List<StudyCountLocalDateResponse> result = List.of();
-
-        BDDMockito.given(studySessionService.getStudyCount(2024, 9, 1L))
-                .willReturn(result);
-
-        // when // then
-        mockMvc.perform(
-                        get("/api/v1/study/count").with(csrf())
-                                .param("year", "2024")
-                                .param("month", "9")
-                                .contentType(MediaType.APPLICATION_JSON)
-                )
-                .andDo(print())
-                .andExpect(status().isOk());
-    }
-
-    @DisplayName("현재 날짜 기준으로 주단위 월, 일, 공부 횟수를 가져옵니다.")
-    @Test
-    void getStudyCountOfWeek() throws Exception {
-        // given
-        List<StudyCountResponse> result = List.of();
-
-        BDDMockito.given(studySessionService.getStudyCountOfWeek(LocalDate.now(), 1L))
-                .willReturn(result);
-
-        // when // then
-        mockMvc.perform(
-                        get("/api/v1/study/week").with(csrf())
-                                .contentType(MediaType.APPLICATION_JSON)
-                )
-                .andDo(print())
-                .andExpect(status().isOk());
-    }
-
-    @DisplayName("yearWeek를 파라미터로 해당 주의 랭킹을 구합니다.")
-    @Test
-    void getStudyDurationRanking() throws Exception {
-        // given
-        final int yearWeek = 202404;
-        WeeklyRankingResponse result = WeeklyRankingResponse.builder().build();
-
-        BDDMockito.given(studySessionService.getStudyDurationRanking(yearWeek))
-                .willReturn(result);
-
-        // when // then
-        mockMvc.perform(
-                        get("/api/v1/study/ranking").with(csrf())
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .param("yearWeek", String.valueOf(yearWeek))
-                )
-                .andDo(print())
-                .andExpect(status().isOk());
     }
 }
